+++ react/components/SKUSelector/Wrapper.js
@@ -1,34 +1,29 @@
 import React, { useContext } from 'react'
 import { ProductContext } from 'vtex.product-context'
-import { path, isEmpty, pathOr } from 'ramda'
+import { isEmpty, pathOr } from 'ramda'
 
-import SKUSelector from './index';
+import SKUSelector from './index'
 
-const SKUSelectorWrapper = (props) => {
+const SKUSelectorWrapper = props => {
   const valuesFromContext = useContext(ProductContext)
 
-  const skuSelectorProps = () => {
-    if (!valuesFromContext || isEmpty(valuesFromContext)) {
-      return props
-    }
+  const skuItems =
+    props.skuItems != null
+      ? props.skuItems
+      : pathOr([], ['product', 'items'], valuesFromContext)
 
-    const { product, selectedItem } = valuesFromContext
+  const skuSelected =
+    props.onSKUSelected != null
+      ? props.skuSelected
+      : valuesFromContext.selectedItem
 
-    return {
-      ...props,
-      skuItems: pathOr([], ['items'], product),
-      skuSelected: selectedItem,
-      productSlug: path(['linkText'], product),
-      variations: path(['variations'], selectedItem),
-    }
-  }
-
-  const { variations, ...restProps } = skuSelectorProps()
-
-  if (!variations || !variations.length) return null
-  
   return (
-    <SKUSelector { ...restProps } />
+    <SKUSelector
+      alwaysShowSecondary={props.alwaysShowSecondary}
+      onSKUSelected={props.onSKUSelected}
+      skuItems={skuItems}
+      skuSelected={skuSelected}
+    />
   )
 }