+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Use props as default value instead of context.
+
 ## [3.42.2] - 2019-06-05
 
 ### Fixed